+++ src/main/java/com/funck/caju/transactionauthorizer/domain/services/AccountService.java
@@ -4,7 +4,7 @@ import com.funck.caju.transactionauthorizer.domain.model.Account;
 
 public interface AccountService {
 
-    Account getAccountById(Integer id);
+    Account getAccountById(Long id);
 
     Account save(Account account);
